+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ##### Enhancements
 
+* LeadingWhitespaceRule is now auto correctable.  
+  [masters3d](https://github.com/masters3d)
+
 * Add included regex for custom rules to control what files are processed.  
   [bootstraponline](https://github.com/bootstraponline)
   [#689](https://github.com/realm/SwiftLint/issues/689)